feat: nft mint item component and styles
