+++ src/App.css
@@ -1,0 +1,7 @@
+.gridContainer {
+  padding: 20px
+}
+
+.card-actions {
+  float: right
+}
